Added missing skin states to the ButtonWithError component.
Used constants for state names, instead of strings directly. The constants are accessible publicly, as static constants.

Removed the binding of showValidity and isValid between the DDL and the ButtonWithError components. The only time these DDL variables are changed are when the state of the DDL changes. So, setting the variables on the button in the state change handler. Hence, making these DDL variables private (the public was needed for the binding).

Changed the name of the value setter to selectedLabel.