Added support for controlProperty in CheckBoxGridColumn.
The check box column can be given the name of a boolean property which is set and unset on the data provider items when the checkbox is ticked and unticked.

If no controlProperty is given, then the check box controls the selection of the rows by default.

The check box header now supports showing header text.