Major update: adding "dataProvider" functionality to the component.
Not using "values", "fixedValues" anymore. The position of the thumbs and the labels of the track highlights can be specified using a dataProvider collection. Changes to the dataProvider are automatically reflected in the component and vice versa.

Refactoring all around to avoid using indices of the thumbs vector and trackHighlightButtons vector to find the previous and next components. Because the addition and removal of the thumbs / track highlights can happen in any order, the above vectors cannot be relied upon to have the items in increasing-value order.