Building a new component that displays a hierarchical dataProvider in the layout of an organizational chart / hierarchical chart.
It is fully functional, except for the visual eye-candy. Currently, the component has only been tested with a dataProvider having children under a property called "children" (which is the default format for hierarchical dataProviders, as accepted by the default tree data descriptor. A custom dataDescriptor can be assigned, but I haven't tested that.